% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_community_index.R
\name{calc_community_index}
\alias{calc_community_index}
\title{Calculate Community Index}
\usage{
calc_community_index(dat_niche, dat_community)
}
\arguments{
\item{dat_niche}{A data frame that contains the species climate niche.}

\item{dat_community}{A list containing community data from observations and experiments.}
}
\value{
A list with calculated community indices for observations and experiments.
}
\description{
This function calculates community indices including CTI and CPI, by taking the community weighted mean of species climate niches.
}
\examples{
\dontrun{
 dat_index <- calc_community_index(dat_niche = dat_niche, dat_community = dat_community)
 dat_index$obs
 dat_index$exp
}
}
