% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_community_shift.R
\name{calc_community_shift}
\alias{calc_community_shift}
\title{Calculate Community Shift}
\usage{
calc_community_shift(dat_index)
}
\arguments{
\item{dat_index}{A list containing community indices CTI and CPI in observations and experiments.}
}
\value{
A list with calculated community shifts in observations and experiments.
}
\description{
This function calculates community shift from community indices CTI and CPI.
}
\examples{
\dontrun{
 dat_shift <- calc_community_shift(dat_index = dat_index)
}
}
