% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_species_gainloss.R
\name{calc_species_gainloss}
\alias{calc_species_gainloss}
\title{Calculate Species Abundance Change}
\usage{
calc_species_gainloss(dat_community, dat_niche)
}
\arguments{
\item{dat_community}{A list containing community data from observations and experiments.}

\item{dat_niche}{A data frame that contains the species climate niche.}
}
\value{
A list with calculated species abundance change in observations and experiments.
}
\description{
This function calculates species abundance change (gain and loss) in communities over time or under warming treatments.
}
\examples{
\dontrun{
  dat_gainloss <- calc_species_gainloss(dat_community = dat_community, dat_niche = dat_niche)
}
}
