% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_species_niche.R
\name{calc_species_niche}
\alias{calc_species_niche}
\title{Calculate Species Niche}
\usage{
calc_species_niche(dat_trait, add_dummy = T)
}
\arguments{
\item{dat_trait}{A data frame that contains climatic conditions associated with individuals (i.e., traits in the climate niche space).}

\item{add_dummy}{A logical argument that determines whether to add dummy species to the output.
If `TRUE`, it adds `Avena DUMMY`, `Festuca DUMMY` and `Hypochaeris DUMMY` to the species niche data frame,
with their niche values calculated as the average of specific species within their respective genera. Default is `TRUE`.}
}
\value{
A data frame with species as rows and calculated niche statistics as columns.
}
\description{
This function calculates a range of statistics to characterize species climate niche,
such as mean, standard deviation, median, and quantiles for the temperature, precipitation, and vapor pressure deficit.
Optionally, this function can also calculate these statistics for the climatic water deficit (cwd) if provided in the data.
}
\examples{
\dontrun{
  species_niche <- calc_species_niche(dat_trait, add_dummy = TRUE)
}
}
