% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_community_index.R
\name{plot_community_index}
\alias{plot_community_index}
\title{Plot Community Index}
\usage{
plot_community_index(
  option,
  dat_index,
  experiment = NULL,
  treatment = NULL,
  layout = "surround",
  nrow = NULL,
  onesite = NULL
)
}
\arguments{
\item{option}{A string to choose between two types of plots. It should be set to "obs" to plot changes over time at observational sites, or set of "exp" to plot changes under treatments in global change experiments.}

\item{dat_index}{A list with calculated community indices for observations and experiments.}

\item{experiment}{A string indicating the specific experiment to plot for. Only valid when "option" is set to "exp". Options are "jrgce" for the Jasper Ridge Global Change Experiment, "mclexp" for the McLaughlin Water Experiment, and "scide" for the Santa Cruz International Drought Experiment.}

\item{treatment}{A string indicating the specific treatment in the experiment to plot for. Only valid when "option" is set to "exp" and "experiment" is set tp "jrgce". Options are "warming" or "watering".}

\item{layout}{A string indicating the layout of the observational plot, either "surround" to have all panels surrounding the site map, or "landsc" to have a simple landscape layout. Only valid when "option" is set to "obs".}

\item{nrow}{An integer indicating the number of rows in the observational plot. Only valid when "option" is set to "obs" and "layout" is set to "landsc".}

\item{onesite}{An optional parameter specifying a specific site to be plotted.}
}
\value{
A ggplot2 object of community index plot.
}
\description{
This function plots the changes in community indices including CTI and CPI over time at observational sites or under treatments in global change experiments.
}
\examples{
\dontrun{
plot_community_index(option = "obs", dat_index, layout = "surround")
plot_community_index(option = "exp", dat_index, experiment = "jrgce", treatment = "warming")
}
}
