% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_community_shift.R
\name{plot_community_shift}
\alias{plot_community_shift}
\title{Plot Community Shift}
\usage{
plot_community_shift(dat_shift, dat_niche)
}
\arguments{
\item{dat_shift}{A list that contains community shift data.}

\item{dat_niche}{A data frame containing species climate niche.}
}
\value{
A list containing ggplot objects of different plots, including community shifts across observational sites, community shifts in the JRGCE warming experiment, comparison of community shifts between observations and experiments, communities in the climate niche space, and a combined plot.
}
\description{
This function creates plots showing community shifts measured by CTI and CPI in the climate niche space across observational sites and in the JRGCE warming experiment.
}
\examples{
\dontrun{
  p_community_shift <- plot_community_shift(dat_shift, dat_niche)
}
}
