% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_effect_size.R
\name{plot_effect_size}
\alias{plot_effect_size}
\title{Plot Effect Size}
\usage{
plot_effect_size(df_index_change_exp)
}
\arguments{
\item{df_index_change_exp}{A data frame containing the experimental data,
with expected columns such as 'trt' for treatment, 'estimate' for effect size,
'std.error' for the standard error, 'exp' for experiment identifier, 'grp'
for group identifier, 'sig' for significance flag, and 'index' for the type
of index being measured (e.g., temperature or precipitation index).}
}
\value{
A ggplot2 object of a coefficient plot that visualizes the effects of experimental manipulations on community indices in multiple global change experiments.
}
\description{
This function visualizes the estimated effects of different experimental manipulations on community indices like CTI and CPI.
}
\examples{
\dontrun{
df_index_change_exp <- test_index_change_all(dat_index, option = "exp")
gg_effect_size <- plot_effect_size(df_index_change_exp)
}
}
