% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_individual_trait_species_niche.R
\name{plot_individual_trait_species_niche_all}
\alias{plot_individual_trait_species_niche_all}
\title{Plot Species Niche Estimated from Individual Traits (All Species)}
\usage{
plot_individual_trait_species_niche_all(
  dat_occ,
  dat_trait,
  dat_niche,
  cool_species = "Danthonia californica",
  warm_species = "Stipa pulchra",
  frac = 1
)
}
\arguments{
\item{dat_occ}{A data frame that contains species occurrence data.}

\item{dat_trait}{A data frame that contains climatic conditions associated with individuals (i.e., traits in the climate niche space).}

\item{dat_niche}{A data frame that contains the species climate niche.}

\item{cool_species}{A string representing a species associated with cool conditions (defaulted to "Danthonia californica").}

\item{warm_species}{A string representing a species associated with warm conditions (defaulted to "Stipa pulchra").}

\item{frac}{A numeric value that represents the fraction of the data to be used in plotting. The default value is 1 which means all data will be used. Warning: A value of 1 will make the figure very slow to render.}
}
\value{
A list of ggplot2 objects of distribution of individuals in the geographical space, distribution of individuals in the climate niche, climate niche of all species, and a combined plot.
}
\description{
This function plots the distribution of all species and their individuals in the geographical and climate space.
Blue and red colors highlight two species associated with cool and warm conditions, respectively.
}
\examples{
\dontrun{
p_individual_trait_species_niche_all <- plot_individual_trait_species_niche_all(dat_occ = dat_occ, dat_trait = dat_trait, dat_niche = dat_niche, frac = 0.01)
}
}
