% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_individual_trait_species_niche.R
\name{plot_individual_trait_species_niche_ind_sp}
\alias{plot_individual_trait_species_niche_ind_sp}
\title{Plot Species Niche Estimated from Individual Traits (One Specific Species)}
\usage{
plot_individual_trait_species_niche_ind_sp(
  dat_occ,
  dat_trait,
  sf_cfp = NULL,
  sp = "Danthonia californica"
)
}
\arguments{
\item{dat_occ}{A data frame that contains species occurrence data.}

\item{dat_trait}{A data frame that contains climatic conditions associated with individuals (i.e., traits in the climate niche space).}

\item{sf_cfp}{A simple feature object that contains the California Floristic Province data.
By default, it reads the "cfp" file from the package's "extdata" directory.}

\item{sp}{A string that represents a specific species of focus (defaulted to "Danthonia californica").}
}
\value{
A combined ggplot2 object that shows the distributions of individuals in the geographical and climate space, leafing to the estimation of species climate niche.
}
\description{
This function plots the distribution of a species and its individuals in the geographical and climate space.
}
\examples{
\dontrun{
  p_individual_trait_species_niche_ind_sp <- plot_individual_trait_species_niche_ind_sp(dat_occ = dat_occ, dat_trait = dat_trait, sp = "Danthonia californica")
}
}
