% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_site_map.R
\name{plot_site_map}
\alias{plot_site_map}
\title{Plot Site Map}
\usage{
plot_site_map(sf_cfp = NULL, ras_grass = NULL, plotsite = T)
}
\arguments{
\item{sf_cfp}{A simple feature object that contains the California Floristic Province data.
By default, it reads the "cfp" file from the package's "extdata" directory.}

\item{ras_grass}{A raster object that contains the grassland percent cover data.
By default, it reads the "cfp-grassland-percent-cover.tif" file from the package's "extdata" directory.}

\item{plotsite}{A boolean variable that decides whether to include sites in the plot. Default is TRUE.}
}
\value{
A ggplot object of the site map.
}
\description{
This function reads and plots a map showing the grassland percentage cover,
and optionally plots the sites with observed data.
}
\examples{
\dontrun{
site_map <- plot_site_map()
site_map
}
}
