% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_species_gainloss.R
\name{plot_species_gainloss}
\alias{plot_species_gainloss}
\title{Plot Species Abundance Change}
\usage{
plot_species_gainloss(dat_niche, dat_gainloss, nrow = 3)
}
\arguments{
\item{dat_niche}{A data frame that contains the species climate niche.}

\item{dat_gainloss}{A list containing species abundance change in observations and experiments.}

\item{nrow}{Numeric determining the number of rows for the layout in the observation plot. Default is 3.}
}
\value{
A list of ggplot objects for the species abundance change at observational sites, in the experiment, and a combined plots.
}
\description{
This function generates plots of species abundance change (gain and loss) in communities over time or under warming treatments.
}
\examples{
\dontrun{
 p_gainloss <- plot_species_gainloss(dat_niche = dat_niche, dat_gainloss = dat_gainloss)
 p_gainloss$obs
 p_gainloss$exp
 p_gainloss$combined
}
}
