% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_species_niche.R
\name{plot_species_niche}
\alias{plot_species_niche}
\title{Plot Species Niche}
\usage{
plot_species_niche(
  dat_niche,
  cool_species = "Danthonia californica",
  warm_species = "Stipa pulchra"
)
}
\arguments{
\item{dat_niche}{A data frame that contains the species climate niche.}

\item{cool_species}{A string representing a species associated with cool conditions (defaulted to "Danthonia californica").}

\item{warm_species}{A string representing a species associated with warm conditions (defaulted to "Stipa pulchra").}
}
\value{
A ggplot2 object showing species climate niche.
}
\description{
This function plots species in the climate niche space based on their temperature and precipitation optima.
Blue and red colors highlight two species associated with cool and warm conditions, respectively.
}
\examples{
\dontrun{
  p_niche <- plot_species_niche(dat_niche = my_data)
  p_niche <- plot_species_niche(dat_niche = my_data, cool_species = "Cool Species", warm_species = "Warm Species")
}
}
