% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_site_info.R
\name{read_site_info}
\alias{read_site_info}
\title{Import Site Information}
\usage{
read_site_info(subset = NULL)
}
\arguments{
\item{subset}{A character value that suggests if a subset of the data is needed.
If `subset` equals "obs", the function returns information of only observational sites.}
}
\value{
A spatial tibble (sf object) with site information.
}
\description{
This function imports site information data and optionally subsets it to only observational sites.
}
\examples{
\dontrun{
  site_info <- read_site_info()
  site_info <- read_site_info(subset = "obs")
}
}
