% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_warm_treatment.R
\name{read_warm_treatment}
\alias{read_warm_treatment}
\title{Import Warming Treatment Data}
\usage{
read_warm_treatment()
}
\value{
A tibble data frame with columns:
\itemize{
  \item `tag`: numbered tag for each phase.
  \item `phase`: the phase designation (e.g., "Phase I").
  \item `name`: the full name and description of each phase.
  \item `start`: start year of each phase (set to Inf for the first phase).
  \item `end`: end year of each phase (set to Inf for the last phase).
  \item `startyear`: start year of each phase.
}
}
\description{
This function imports the warming treatments in the Jasper Ridge Global Change Experiment.
The warming manipulations consist of three phases with different treatments.
}
\examples{
\dontrun{
  treatments <- read_warm_treatment()
}
}
