% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_change.R
\name{test_index_change_all}
\alias{test_index_change_all}
\title{Test Community Index Change in All Groups}
\usage{
test_index_change_all(dat_index, option)
}
\arguments{
\item{dat_index}{A list with calculated community indices for observations and experiments.}

\item{option}{A string to choose between two types of tests. It should be set to "obs" to test changes over time at observational sites, or set of "exp" to test changes under treatments in global change experiments.}
}
\value{
A data frame with the magnitude, uncertainty, and significance level of changes in community index.
}
\description{
This function statistically tests changes in community indices across all available groups.
}
\examples{
\dontrun{
df_index_change_obs <- test_index_change_all(dat_index = dat_index, option = "obs")
df_index_change_obs
df_index_change_exp <- test_index_change_all(dat_index = dat_index, option = "exp")
df_index_change_exp
}
}
