% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_change.R
\name{test_optima_change_all}
\alias{test_optima_change_all}
\title{Test Species Optima Change in All Groups}
\usage{
test_optima_change_all(dat_community, dat_niche, option)
}
\arguments{
\item{dat_community}{A list containing community data from observations and experiments.}

\item{dat_niche}{A data frame that contains the species climate niche.}

\item{option}{A string to choose between two types of tests. It should be set to "obs" to test changes over time at observational sites, or set of "exp" to test changes under treatments in global change experiments.}
}
\value{
A data frame with the magnitude, uncertainty, and significance level of changes in species optima.
}
\description{
This function statistically tests changes in species optima across all available groups.
}
\examples{
\dontrun{
  df_optima_change_obs <- test_optima_change_all(dat_community = dat_community, dat_niche = dat_niche, option = "obs")
  df_optima_change_obs
  df_optima_change_exp <- test_optima_change_all(dat_community = dat_community, dat_niche = dat_niche, option = "exp")
  df_optima_change_exp
}
}
